requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Legacy Flask (if needed)
Flask==2.3.3
flask-pymongo==2.3.0
flask-wtf==1.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-mail==0.9.1

# AI Services
openai==1.3.5
groq>=0.4.0
# Core Dependencies
itsdangerous==2.1.2
werkzeug==2.3.7
wtforms==3.0.1
pymongo==4.5.0
python-dotenv
bcrypt==4.0.1

# CORS
flask-cors
fastapi-cors

# Google AI
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.10
# PDF/Document Processing - Modern Stack
pypdf>=4.0.0  # Modern PDF extraction (replaces pdfminer)
PyMuPDF==1.22.5  # Layout-preserving extraction
python-docx==1.1.0
pdfplumber==0.7.6  # Legacy fallback
pdfminer.six==20221105  # Legacy fallback

# NLP & Text Processing
spacy==3.7.2
phonenumbers==8.13.26
nltk>=3.8.1

# Modern Data Validation & Formatting
pydantic>=2.5.0  # Structured data validation
rich>=13.7.0  # Beautiful terminal formatting and HTML export

# OCR Support for scanned PDFs
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.16.3

# Additional CV Analysis Libraries
pdfminer3>=2018.12.3.0
pafy>=0.5.5
youtube-dl>=2021.12.17

# Testing
pytest>=8.0.0
reportlab>=4.0.0  # Modern PDF generation
xhtml2pdf>=0.2.13  # HTML-to-PDF conversion
cssselect2>=0.7.0

# Professional PDF Generation (Windows-compatible)
xhtml2pdf>=0.2.13